Move procedure name functions to internal/procedure
